implemented selective rebuild for zsync files

+++ kekosync-server/src/main/java/com/kellerkompanie/kekosync/server/helper/ZsyncGenerator.java
@@ -1,12 +1,15 @@
 package com.kellerkompanie.kekosync.server.helper;
 
 import com.salesforce.zsync.ZsyncMake;
+import com.salesforce.zsync.internal.ControlFile;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
 import java.io.IOException;
+import java.io.InputStream;
 import java.nio.file.*;
+import java.util.Date;
 
 import static com.kellerkompanie.kekosync.server.constants.FileMatcher.sourceFileMatcher;
 import static com.kellerkompanie.kekosync.server.constants.FileMatcher.zsyncFileMatcher;
@@ -28,11 +31,12 @@ public class ZsyncGenerator {
     }
 
     private static void processFile(Path sourceFilePath) {
-        //Path zsyncFilePath = zsyncMake.make(sourceFilePath);
+        if (Files.exists(getPathOfZsync(sourceFilePath))) return;
+
         ZsyncMake.Options options = new ZsyncMake.Options();
         options.setBlockSize(8192); //trying to stay compatible with lé arma3sync.
         Path zsyncFilePath = zsyncMake.writeToFile(sourceFilePath, options).getOutputFile();
-        log.debug("{} -> {}", sourceFilePath, zsyncFilePath);
+        log.info("{} -> {}", sourceFilePath, zsyncFilePath);
     }
 
     public static void cleanDirectory(String directoryPath) throws IOException {
@@ -44,11 +48,34 @@ public class ZsyncGenerator {
 
     private static void deleteFile(Path filePath) {
         try {
-            Files.delete(filePath);
-            log.debug("deleted {}", filePath);
+            InputStream inputStream = Files.newInputStream(filePath);
+            ControlFile controlFile = ControlFile.read(inputStream);
+            inputStream.close();
+
+            long zsyncMTime = controlFile.getHeader().getMtime().getTime();
+            long zsyncSize = controlFile.getHeader().getLength();
+            Path originalFile = getPathOfOriginal(filePath);
+            long originalMTime = (Files.getLastModifiedTime(originalFile).toMillis() / 1000) * 1000;  // zsync files are ignoring millisecond precision
+            long originalSize = Files.size(originalFile);
+
+            // only delete if zsync is outdated
+            if ((zsyncMTime < originalMTime) || zsyncSize != originalSize) {
+                Files.delete(filePath);
+                log.info("deleted {}", filePath);
+            }
         } catch (IOException e) {
             log.error("failed to delete {}", filePath, e);
         }
+    }
+
+    private static Path getPathOfOriginal(Path zsyncFilepath) {
+        String fileName = zsyncFilepath.getFileName().toString();
+        String originalFileName = fileName.substring(0, fileName.lastIndexOf("."));
+        return zsyncFilepath.resolveSibling(originalFileName);
+    }
+
+    private static Path getPathOfZsync(Path sourceFilePath) {
+        return sourceFilePath.resolveSibling(sourceFilePath.getFileName() + ".zsync");
     }
 
 }
